+++ src/extension/figure/rect.ts
@@ -60,39 +60,40 @@ export function drawRect (ctx: CanvasRenderingContext2D, attrs: RectAttrs | Rect
     borderRadius: r = 0,
     borderDashedValue = [2, 2]
   } = styles
-  if (
-    (style === PolygonType.Fill || styles.style === PolygonType.StrokeFill) &&
-    (!isString(color) || !isTransparent(color))
-  ) {
+  // eslint-disable-next-line @typescript-eslint/unbound-method
+  const draw = ctx.roundRect ?? ctx.rect
+  const solid = (style === PolygonType.Fill || styles.style === PolygonType.StrokeFill) && (!isString(color) || !isTransparent(color))
+  if (solid) {
     ctx.fillStyle = color
     rects.forEach(({ x, y, width: w, height: h }) => {
       ctx.beginPath()
-      ctx.moveTo(x + r, y)
-      ctx.arcTo(x + w, y, x + w, y + h, r)
-      ctx.arcTo(x + w, y + h, x, y + h, r)
-      ctx.arcTo(x, y + h, x, y, r)
-      ctx.arcTo(x, y, x + w, y, r)
+      draw.call(ctx, x, y, w, h, r)
       ctx.closePath()
       ctx.fill()
     })
   }
   if ((style === PolygonType.Stroke || styles.style === PolygonType.StrokeFill) && borderSize > 0 && !isTransparent(borderColor)) {
     ctx.strokeStyle = borderColor
+    ctx.fillStyle = borderColor
     ctx.lineWidth = borderSize
     if (borderStyle === LineType.Dashed) {
       ctx.setLineDash(borderDashedValue)
     } else {
       ctx.setLineDash([])
     }
+    const correction = borderSize % 2 === 1 ? 0.5 : 0
+    const doubleCorrection = Math.round(correction * 2)
     rects.forEach(({ x, y, width: w, height: h }) => {
-      ctx.beginPath()
-      ctx.moveTo(x + r, y)
-      ctx.arcTo(x + w, y, x + w, y + h, r)
-      ctx.arcTo(x + w, y + h, x, y + h, r)
-      ctx.arcTo(x, y + h, x, y, r)
-      ctx.arcTo(x, y, x + w, y, r)
-      ctx.closePath()
-      ctx.stroke()
+      if (w > borderSize * 2 && h > borderSize * 2) {
+        ctx.beginPath()
+        draw.call(ctx, x + correction, y + correction, w - doubleCorrection, h - doubleCorrection, r)
+        ctx.closePath()
+        ctx.stroke()
+      } else {
+        if (!solid) {
+          ctx.fillRect(x, y, w, h)
+        }
+      }
     })
   }
 }