+++ src/view/OverlayView.ts
@@ -17,7 +17,6 @@ import type Coordinate from '../common/Coordinate'
 import type Point from '../common/Point'
 import type Bounding from '../common/Bounding'
 import type BarSpace from '../common/BarSpace'
-import type Precision from '../common/Precision'
 import { type OverlayStyle } from '../common/Styles'
 import { type EventHandler, type EventName, type MouseTouchEvent, type MouseTouchEventCallback } from '../common/SyntheticEvent'
 import { isBoolean, isNumber, isValid } from '../common/utils/typeChecks'
@@ -27,7 +26,7 @@ import { type CustomApi } from '../Options'
 import type Axis from '../component/Axis'
 import type XAxis from '../component/XAxis'
 import type YAxis from '../component/YAxis'
-import { type OverlayFigure, type OverlayFigureIgnoreEventType } from '../component/Overlay'
+import { type OverlayPrecision, type OverlayFigure, type OverlayFigureIgnoreEventType } from '../component/Overlay'
 import type Overlay from '../component/Overlay'
 import { OVERLAY_FIGURE_KEY_PREFIX, OverlayMode, getAllOverlayFigureIgnoreEventTypes } from '../component/Overlay'
 
@@ -393,10 +392,26 @@ export default class OverlayView<C extends Axis = YAxis> extends View<C> {
     const hoverInstanceInfo = overlayStore.getHoverInstanceInfo()
     const clickInstanceInfo = overlayStore.getClickInstanceInfo()
     const overlays = this.getCompleteOverlays(overlayStore, paneId)
+    const paneIndicators = chartStore.getIndicatorStore().getInstances(paneId)
+    const overlayPrecision = paneIndicators.reduce((prev, indicator) => {
+      const precision = indicator.precision
+      prev[indicator.name] = precision
+      prev.max = Math.max(prev.max, precision)
+      prev.min = Math.min(prev.min, precision)
+      prev.excludePriceVolumeMax = Math.max(prev.excludePriceVolumeMax, precision)
+      prev.excludePriceVolumeMin = Math.min(prev.excludePriceVolumeMin, precision)
+      return prev
+    }, {
+      ...precision,
+      max: Math.max(precision.price, precision.volume),
+      min: Math.min(precision.price, precision.volume),
+      excludePriceVolumeMax: Number.MIN_SAFE_INTEGER,
+      excludePriceVolumeMin: Number.MAX_SAFE_INTEGER
+    })
     overlays.forEach(overlay => {
       if (overlay.visible) {
         this._drawOverlay(
-          ctx, overlay, bounding, barSpace, precision,
+          ctx, overlay, bounding, barSpace, overlayPrecision,
           dateTimeFormat, customApi, thousandsSeparator, decimalFoldThreshold,
           defaultStyles, xAxis, yAxis,
           hoverInstanceInfo, clickInstanceInfo, timeScaleStore
@@ -410,7 +425,7 @@ export default class OverlayView<C extends Axis = YAxis> extends View<C> {
       if (overlay !== null && overlay.visible) {
         this._drawOverlay(
           ctx, overlay, bounding, barSpace,
-          precision, dateTimeFormat, customApi, thousandsSeparator, decimalFoldThreshold,
+          overlayPrecision, dateTimeFormat, customApi, thousandsSeparator, decimalFoldThreshold,
           defaultStyles, xAxis, yAxis,
           hoverInstanceInfo, clickInstanceInfo, timeScaleStore
         )
@@ -423,7 +438,7 @@ export default class OverlayView<C extends Axis = YAxis> extends View<C> {
     overlay: Overlay,
     bounding: Bounding,
     barSpace: BarSpace,
-    precision: Precision,
+    precision: OverlayPrecision,
     dateTimeFormat: Intl.DateTimeFormat,
     customApi: CustomApi,
     thousandsSeparator: string,
@@ -512,7 +527,7 @@ export default class OverlayView<C extends Axis = YAxis> extends View<C> {
     coordinates: Coordinate[],
     bounding: Bounding,
     barSpace: BarSpace,
-    precision: Precision,
+    precision: OverlayPrecision,
     thousandsSeparator: string,
     decimalFoldThreshold: number,
     dateTimeFormat: Intl.DateTimeFormat,
@@ -528,7 +543,7 @@ export default class OverlayView<C extends Axis = YAxis> extends View<C> {
     overlay: Overlay,
     coordinates: Coordinate[],
     _bounding: Bounding,
-    _precision: Precision,
+    _precision: OverlayPrecision,
     _dateTimeFormat: Intl.DateTimeFormat,
     _customApi: CustomApi,
     _thousandsSeparator: string,